scangobj: fix compil warning [-Wdouble-promotion]
- describe_double_constant() takes a gdouble param.
- g_ascii_formatd() takes as last param a gdouble.

When it's a GParamSpecFloat, cast the individual fields (that are
gfloat) to gdouble where a gdouble is expected.

Running GTK-Doc with -Wdouble-promotion generated these kind of
warnings:
```
Building documentation for libgedit-gfls-1
libgedit-gfls-1-scan.c:904:46: warning: implicit conversion increases floating-point precision: 'gfloat' (aka 'float') to 'gdouble' (aka 'double') [-Wdouble-promotion]
  904 |     lower = describe_double_constant (pspec->minimum);
      |             ~~~~~~~~~~~~~~~~~~~~~~~~  ~~~~~~~^~~~~~~
libgedit-gfls-1-scan.c:905:46: warning: implicit conversion increases floating-point precision: 'gfloat' (aka 'float') to 'gdouble' (aka 'double') [-Wdouble-promotion]
  905 |     upper = describe_double_constant (pspec->maximum);
      |             ~~~~~~~~~~~~~~~~~~~~~~~~  ~~~~~~~^~~~~~~
libgedit-gfls-1-scan.c:1066:16: warning: implicit conversion increases floating-point precision: 'gfloat' (aka 'float') to 'gdouble' (aka 'double') [-Wdouble-promotion]
 1065 |     g_ascii_formatd (desc, G_ASCII_DTOSTR_BUF_SIZE, "%g",
      |     ~~~~~~~~~~~~~~~
 1066 |         pspec->default_value);
      |         ~~~~~~~^~~~~~~~~~~~~
3 warnings generated.
```

This is now fixed.

+++ gtkdoc/scangobj.py
@@ -939,8 +939,8 @@ describe_type (GParamSpec *spec)
   else if (G_IS_PARAM_SPEC_FLOAT (spec)) {
     GParamSpecFloat *pspec = G_PARAM_SPEC_FLOAT (spec);
 
-    lower = describe_double_constant (pspec->minimum);
-    upper = describe_double_constant (pspec->maximum);
+    lower = describe_double_constant ((gdouble) pspec->minimum);
+    upper = describe_double_constant ((gdouble) pspec->maximum);
     if (GTKDOC_COMPARE_FLOAT (pspec->minimum, -G_MAXFLOAT)) {
       if (GTKDOC_COMPARE_FLOAT (pspec->maximum, G_MAXFLOAT))
         desc = g_strdup ("");
@@ -1101,7 +1101,7 @@ describe_default (GParamSpec *spec)
      * and do not need the exact same bit representation when deserialising */
     desc = g_malloc0 (G_ASCII_DTOSTR_BUF_SIZE);
     g_ascii_formatd (desc, G_ASCII_DTOSTR_BUF_SIZE, "%g",
-        pspec->default_value);
+        (gdouble) pspec->default_value);
   }
   else if (G_IS_PARAM_SPEC_DOUBLE (spec)) {
     GParamSpecDouble *pspec = G_PARAM_SPEC_DOUBLE (spec);
